+++ adminer/drivers/mssql.inc.php
@@ -441,8 +441,8 @@ WHERE OBJECT_NAME(i.object_id) = " . q($table), $connection2) as $row
 		return $return;
 	}
 
-	function information_schema(?string $db): bool {
-		return get_schema() == "INFORMATION_SCHEMA";
+	function information_schema(?string $db, string $schema = ""): bool {
+		return in_array($schema != "" ? $schema : get_schema(), array("INFORMATION_SCHEMA", "sys"));
 	}
 
 	function error(): string {